Video: Load right LibGL dynamic library
The file `/usr/lib/libgl.so` is a symlink to `libgl.so.serenity`.
Because `__dlopen()` considers the basename to be the library's name,
two sets of memory regions for LibGL could be assigned to software
using SDL.

For example, the `g_gl_context` pointer could be set for `libgl.so` but
then cause a null deref in `libgl.so.serenity`.

Fix this by changing the `dlopen()` call to the right path.