Video: Prevent `GUI::Application` from exiting
The SDL application should be in control. For example, in ScummVM, the
application decides to close the launcher window before opening a new
window for the game when running in OpenGL / software hybrid rendering
mode, resulting in `GUI::Application` calling it quits as soon as the
launcher window is gone.