Events: Keep pumping EventLoop until empty
Events might dispatch other events that we may want to handle
immediately rather than on the next iteration of PumpEvents.

For example, a mouse move event is sent to the Window as an IPC message
which then dispatches a new mouse move event. This means that without
this change, SDL applications only receive mouse move events every
other frame.