Video: Implement shifted key mappings
The `GUI::KeyEvent` we receive contains a specific keycode for shifted
keys, but SDL2 expects us to provide the keycode for the unshifted key
with the same scancode. SDL2 then takes care of handling key shifting
by keeping track of whether or not the Shift key was pressed.

This fixes shifted key entry in e.g. the Quake console.